% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_util.r
\name{dev.reset}
\alias{dev.reset}
\title{Reset device parameters}
\usage{
dev.reset(dev = dev.cur(), ...)
}
\arguments{
\item{dev}{numeric Device ID. Defaults to currently active device. If another valid device this will be activated. If not an existing device a new device will be opened.}

\item{...}{Named list with any other valid graphics parameters to apply}
}
\value{
Updated parameters of the active device
}
\description{
Reset some display parameters of the active device, or make one active while setting its parameters
}
\seealso{
par
}
