% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs.r
\name{get.run.par}
\alias{get.run.par}
\title{get.run.par
Initalise model run parameters. Note this function is maintained for backward compatibility only}
\usage{
get.run.par(tms = NULL, dt = NULL, units = "secs", ...)
}
\arguments{
\item{tms}{xts time series or an object that has a POSIXct index}

\item{dt}{Numeric Time step in seconds}

\item{units}{string  Units for dt.}

\item{...}{Any other parameters returned by get.run.par}
}
\value{
Structure to maintain run information.
}
\description{
get.run.par
Initalise model run parameters. Note this function is maintained for backward compatibility only
}
\details{
The returned value includes a simulation times calculated from the supplied time range and interval
}
