% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{Mode}
\alias{Mode}
\title{Mode - modal value}
\usage{
Mode(x)
}
\arguments{
\item{x}{Numeric. A vector of numerical values whose mode is wanted}
}
\value{
The modal value
}
\description{
Mode - modal value
}
\note{
Capitalisation is to distinguish method name from base::mode
}
