% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{new_guid}
\alias{new_guid}
\title{Get one or more GUIDs}
\usage{
new_guid(n = 1, sep = "-", max = 1e+05)
}
\arguments{
\item{n}{numeric Number of guids to return}

\item{sep}{character Separator character}

\item{max}{numeric Maximum for randomised element (default 1e6)}
}
\value{
A new GUID based on the system time + plus a random number in 0 to max
}
\description{
A GUID (globally unique ID) can be used to identify objects such as files uniquely on the system. This method is based on the system time + plus a random number sampled from a uniform distribution
}
