% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_util.r
\name{time_at_peak}
\alias{time_at_peak}
\title{Time of maximum observation}
\usage{
time_at_peak(ts, icol = 1)
}
\arguments{
\item{ts}{Time series}

\item{icol}{Column index if a multi-column time series}
}
\description{
Determine the time of the maximum item in the supplied time series.
}
\examples{

require(dynatopmodel)

data(brompton)

with(brompton$storm.run, time_at_peak(qsim))

}
\author{
Peter Metcalfe

Peter Metcalfe
}

